requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==1.10.15
python-dotenv==1.0.1
SQLAlchemy==2.0.36
asyncpg==0.29.0
redis==5.0.8
httpx==0.27.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
pytest==8.3.3
pytest-asyncio==0.23.8
pytest-cov==5.0.0
requests-cache==1.2.1
starlette==0.38.5
beautifulsoup4==4.12.2
lxml==4.9.3
